+++ tests/StaticTests.js
@@ -57,9 +57,10 @@ fluid.defaults("kettle.tests.middleware.verifyingUnmarked", {
 
 var infusionPackage = fluid.require("%infusion/package.json");
 
-//------------- Test defs for GET, POST, PUT ---------------
+/** Test defs for static hosting **/
+
 kettle.tests["static"].testDefs = [{
-    name: "HTTPMethods GET test",
+    name: "Static hosting tests",
     expect: 9,
     config: {
         configName: "kettle.tests.static.config",